+++ src/winusb/messenger-winusb.cpp
@@ -136,7 +136,7 @@ namespace librealsense
             auto h = _handle->get_interface_handle(in);
 
             auto buffer = const_cast<uint8_t*>(request->get_buffer().data());
-            int res = WinUsb_ReadPipe(h, epa, buffer, request->get_buffer().size(), &read_pipe_transfer_size, ovl);
+            int res = WinUsb_ReadPipe(h, epa, buffer, static_cast<ULONG>(request->get_buffer().size()), &read_pipe_transfer_size, ovl);
             if (0 != res)
                 return winusb_status_to_rs(res);
